+++ crates/cairo-profiler/src/sierra_loader.rs
@@ -1,16 +1,18 @@
-use crate::trace_reader::function_name::FunctionName;
 use anyhow::{anyhow, Context, Result};
+use cairo_annotations::annotations::profiler::{
+    ProfilerAnnotationsV1, VersionedProfilerAnnotations,
+};
+use cairo_annotations::annotations::TryFromDebugInfo;
+use cairo_annotations::trace_data::{CallTraceNode, CallTraceV1};
 use cairo_lang_sierra::debug_info::DebugInfo;
-use cairo_lang_sierra::program::{Program, ProgramArtifact, StatementIdx, VersionedProgram};
+use cairo_lang_sierra::program::{Program, ProgramArtifact, VersionedProgram};
 use cairo_lang_sierra_to_casm::compiler::{CairoProgramDebugInfo, SierraToCasmConfig};
 use cairo_lang_sierra_to_casm::metadata::calc_metadata;
 use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
 use cairo_lang_starknet_classes::contract_class::ContractClass;
 use camino::{Utf8Path, Utf8PathBuf};
-use serde_json::Value;
 use std::collections::HashMap;
 use std::fs;
-use trace_data::{CallTrace, CallTraceNode};
 
 /// Map with sierra and casm debug info needed for function level profiling.
 /// All paths in the map are absolute paths.
@@ -19,7 +21,7 @@ pub struct CompiledArtifactsCache(HashMap<Utf8PathBuf, CompiledArtifacts>);
 pub struct CompiledArtifacts {
     pub sierra_program: SierraProgram,
     pub casm_debug_info: CairoProgramDebugInfo,
-    pub statements_functions_map: Option<StatementsFunctionsMap>,
+    pub statements_functions_map: Option<ProfilerAnnotationsV1>,
 }
 
 pub enum SierraProgram {
@@ -34,19 +36,6 @@ impl SierraProgram {
                 program
             }
         }
-    }
-}
-
-/// This struct maps sierra statement index to a stack of fully qualified paths of cairo functions
-/// consisting of a function which caused the statement to be generated and all functions that were
-/// inlined or generated along the way, up to the first non-inlined function from the original code.
-/// The map represents the stack from the least meaningful elements.
-#[derive(Default, Clone)]
-pub struct StatementsFunctionsMap(HashMap<StatementIdx, Vec<FunctionName>>);
-
-impl StatementsFunctionsMap {
-    pub fn get(&self, key: StatementIdx) -> Option<&Vec<FunctionName>> {
-        self.0.get(&key)
     }
 }
 
@@ -159,33 +148,15 @@ pub fn compile_sierra_and_add_compiled_artifacts_to_cache(
 
 fn maybe_get_statements_functions_map(
     maybe_sierra_program_debug_info: Option<DebugInfo>,
-) -> Option<StatementsFunctionsMap> {
-    maybe_sierra_program_debug_info.and_then(|mut debug_info| {
-        debug_info
-            .annotations
-            .shift_remove("github.com/software-mansion/cairo-profiler")
-            .map(get_statements_functions_map)
-    })
-}
-
-pub fn get_statements_functions_map(mut annotations: Value) -> StatementsFunctionsMap {
-    assert!(
-        annotations.get("statements_functions").is_some(),
-        "Wrong debug info annotations format"
-    );
-    let statements_functions = annotations["statements_functions"].take();
-    let map = serde_json::from_value::<HashMap<StatementIdx, Vec<String>>>(statements_functions)
-        .expect("Wrong statements function map format");
-
-    StatementsFunctionsMap(
-        map.into_iter()
-            .map(|(key, names)| (key, names.into_iter().map(FunctionName).collect()))
-            .collect(),
-    )
+) -> Option<ProfilerAnnotationsV1> {
+    let VersionedProfilerAnnotations::V1(annotations) =
+        VersionedProfilerAnnotations::try_from_debug_info(&maybe_sierra_program_debug_info?)
+            .ok()?;
+    Some(annotations)
 }
 
 pub fn collect_and_compile_all_sierra_programs(
-    trace: &CallTrace,
+    trace: &CallTraceV1,
 ) -> Result<CompiledArtifactsCache> {
     let mut compiled_artifacts_cache = CompiledArtifactsCache::new();
     collect_compiled_artifacts(trace, &mut compiled_artifacts_cache)?;
@@ -194,7 +165,7 @@ pub fn collect_and_compile_all_sierra_programs(
 }
 
 fn collect_compiled_artifacts(
-    trace: &CallTrace,
+    trace: &CallTraceV1,
     compiled_artifacts_cache: &mut CompiledArtifactsCache,
 ) -> Result<()> {
     if let Some(cairo_execution_info) = &trace.cairo_execution_info {