+++ crates/cairo-profiler/src/trace_reader/function_trace_builder.rs
@@ -1,6 +1,4 @@
 use crate::profiler_config::FunctionLevelConfig;
-use crate::sierra_loader::StatementsFunctionsMap;
-use crate::trace_reader::function_name::FunctionName;
 use crate::trace_reader::function_trace_builder::function_stack_trace::{
     CallStack, VecWithLimitedCapacity,
 };
@@ -10,6 +8,9 @@ use crate::trace_reader::function_trace_builder::stack_trace::{
 };
 use crate::trace_reader::sample::{FunctionCall, InternalFunctionCall, Sample};
 use crate::versioned_constants_reader::OsResources;
+use cairo_annotations::annotations::profiler::{FunctionName, ProfilerAnnotationsV1};
+use cairo_annotations::trace_data::CasmLevelInfo;
+use cairo_annotations::{map_pcs_to_sierra_statement_ids, MappingResult};
 use cairo_lang_sierra::extensions::core::{CoreConcreteLibfunc, CoreLibfunc, CoreType};
 use cairo_lang_sierra::extensions::starknet::StarkNetConcreteLibfunc;
 use cairo_lang_sierra::program::{GenStatement, Program, StatementIdx};
@@ -17,7 +18,6 @@ use cairo_lang_sierra::program_registry::ProgramRegistry;
 use cairo_lang_sierra_to_casm::compiler::CairoProgramDebugInfo;
 use std::collections::HashMap;
 use std::ops::AddAssign;
-use trace_data::TraceEntry;
 
 mod function_stack_trace;
 mod inlining;
@@ -43,37 +43,17 @@ impl AddAssign<usize> for Steps {
     }
 }
 
-enum MaybeSierraStatementIndex {
-    SierraStatementIndex(StatementIdx),
-    PcOutOfFunctionArea,
-}
-
 /// Collects profiling info of the current run using the trace.
 #[allow(clippy::too_many_lines)]
 pub fn collect_function_level_profiling_info(
-    trace: &[TraceEntry],
     program: &Program,
     casm_debug_info: &CairoProgramDebugInfo,
-    run_with_call_header: bool,
-    statements_functions_map: &Option<StatementsFunctionsMap>,
+    casm_level_info: &CasmLevelInfo,
+    statements_functions_map: &Option<ProfilerAnnotationsV1>,
     function_level_config: &FunctionLevelConfig,
     os_resources_map: &OsResources,
 ) -> FunctionLevelProfilingInfo {
     let sierra_program_registry = &ProgramRegistry::<CoreType, CoreLibfunc>::new(program).unwrap();
-
-    // Some CASM programs starts with a header of instructions to wrap the real program.
-    // `real_minimal_pc` is the PC in the trace that points to the same CASM instruction which would
-    // be in the PC=1 in the original CASM program.
-    // This is the same as the PC of the last trace entry plus 1, as the header is built to have
-    // a `ret` last instruction, which must be the last in the trace of any execution.
-    // The first instruction after that is the first instruction in the original CASM program.
-    // This logic only applies when a header was added to the CASM program, otherwise the
-    // `real_minimal_pc` is the default one which is 1.
-    let real_minimal_pc = if run_with_call_header {
-        trace.last().unwrap().pc + 1
-    } else {
-        1
-    };
 
     let mut call_stack = CallStack::new(function_level_config.max_function_stack_trace_depth);
 
@@ -94,30 +74,23 @@ pub fn collect_function_level_profiling_info(
     // That's why we must track entry to a syscall, and leave as soon as we're out of given GenStatement::Invocation
     let mut in_syscall_idx: Option<StatementIdx> = None;
 
-    for step in trace {
-        // Skip the header.
-        if step.pc < real_minimal_pc {
-            header_steps += 1;
-            continue;
-        }
-        // The real pc is equal to (1 + step.pc - real_minimal_pc). This difference however
-        // is a code offset of the real pc. We need real code offset to map pc to sierra statement.
-        let real_pc_code_offset = step.pc - real_minimal_pc;
+    let sierra_statements = map_pcs_to_sierra_statement_ids(casm_debug_info, casm_level_info);
+
+    for statement in sierra_statements {
+        let sierra_statement_idx = match statement {
+            MappingResult::SierraStatementIdx(sierra_statement_idx) => sierra_statement_idx,
+            MappingResult::Header => {
+                header_steps += 1;
+                continue;
+            }
+            MappingResult::PcOutOfFunctionArea => {
+                continue;
+            }
+        };
 
         if end_of_program_reached {
             unreachable!("End of program reached, but trace continues.");
         }
-
-        let maybe_sierra_statement_idx =
-            maybe_sierra_statement_index_by_pc(casm_debug_info, real_pc_code_offset);
-        let sierra_statement_idx = match maybe_sierra_statement_idx {
-            MaybeSierraStatementIndex::SierraStatementIndex(sierra_statement_idx) => {
-                sierra_statement_idx
-            }
-            MaybeSierraStatementIndex::PcOutOfFunctionArea => {
-                continue;
-            }
-        };
 
         let current_function_name = FunctionName::from_sierra_statement_idx(
             sierra_statement_idx,
@@ -175,7 +148,7 @@ pub fn collect_function_level_profiling_info(
                             let mut current_call_stack_with_syscall = current_call_stack.clone();
                             current_call_stack_with_syscall.push(
                                 FunctionCall::InternalFunctionCall(InternalFunctionCall::Syscall(
-                                    FunctionName(map_syscall_to_selector(syscall).into()),
+                                    FunctionName(map_syscall_to_selector(syscall).to_string()),
                                 )),
                             );
                             *syscall_stack_traces
@@ -210,41 +183,12 @@ pub fn collect_function_level_profiling_info(
     }
 }
 
-fn maybe_sierra_statement_index_by_pc(
-    casm_debug_info: &CairoProgramDebugInfo,
-    real_pc_code_offset: usize,
-) -> MaybeSierraStatementIndex {
-    // The `-1` here can't cause an underflow as the first statement's start offset is always 0,
-    // so it is always on the left side of the partition, thus the partition index is > 0.
-    let statement_index = StatementIdx(
-        casm_debug_info
-            .sierra_statement_info
-            .partition_point(|x| x.start_offset <= real_pc_code_offset)
-            - 1,
-    );
-    // End offset is exclusive and the casm debug info is sorted in non-descending order by both
-    // end offset and start offset. Therefore, the end offset of the last element in that vector
-    // is the bytecode length.
-    let bytecode_length = casm_debug_info
-        .sierra_statement_info
-        .last()
-        .unwrap()
-        .end_offset;
-    // If offset is greater or equal the bytecode length it means that it is the outside ret used
-    // for e.g. getting pointer to builtins costs table, const segments etc.
-    if real_pc_code_offset >= bytecode_length {
-        MaybeSierraStatementIndex::PcOutOfFunctionArea
-    } else {
-        MaybeSierraStatementIndex::SierraStatementIndex(statement_index)
-    }
-}
-
 fn build_current_call_stack(
     call_stack: &CallStack,
     current_function_name: FunctionName,
     show_inlined_functions: bool,
     sierra_statement_idx: StatementIdx,
-    statements_functions_map: Option<&StatementsFunctionsMap>,
+    statements_functions_map: Option<&ProfilerAnnotationsV1>,
 ) -> VecWithLimitedCapacity<FunctionCall> {
     let mut current_call_stack = call_stack.current_call_stack().clone();